requirements.txt
flask==3.0.0
line-bot-sdk==3.12.0
google-generativeai==0.8.3
gspread==6.1.2
google-auth==2.35.0
Pillow==10.4.0
gunicorn==23.0.0
